Add BGP announcement

+++ filter/filter.cpp
@@ -25,12 +25,16 @@
 #include "iptable.hpp"
 #include "sqlite.hpp"
 #include "client.hpp"
+#include "bgp.hpp"
 
 #include<ctime>
 #include<cstdlib>
 
 #define END_PHRASE "FINISH"
-bool Sqlite::SQLite::conf =0;
+bool Sqlite::SQLite::conf = false;
+bool Sqlite::SQLite::init_database = false;
+
+using namespace BGP;
 
 void watcher(std::vector<std::shared_ptr<Anomaly>> & threads_collect,
         std::shared_ptr<Anomaly> anomly, std::shared_ptr<ts_queue<token>> task){
@@ -60,20 +64,25 @@ void watcher(std::vector<std::shared_ptr<Anomaly>> & threads_collect,
 }
 bool end_process(std::shared_ptr<Client>& _client, int id, int byte, int packet){
 
+	try{
 	if(!_client->connect())
 		return false;
 	std::string data = std::string(END_PHRASE) + " " + std::to_string(id) + 
 		" " + std::to_string(byte) + " " + std::to_string(packet) + "\n";
-	if(!_client->send(data))
-		return false;
-	if(_client->read("\n") == "OK!\n"){
+	if(_client->send(data))
+		if(_client->read("\n") == "OK!\n"){
+			_client->close();
+			return true;
+		}
+	}catch(...){
 		_client->close();
-		return true;
+		return false;
 	}
+	_client->close();
 	return false;
 }
 void task_runner(std::shared_ptr<ts_queue<token>> task, int id, int timeout,
-	       	uint8_t proto, uint32_t dst_addr, std::shared_ptr<Iptable>& ipt, std::shared_ptr<Client>& _client){
+	       	uint8_t proto, uint32_t dst_addr, std::shared_ptr<Iptable>& ipt, std::shared_ptr<Client>& _client, std::shared_ptr<Bgp>& bgp){
 
     std::chrono::high_resolution_clock::time_point _last = std::chrono::high_resolution_clock::now();
     std::chrono::high_resolution_clock::time_point _now;
@@ -92,9 +101,15 @@ void task_runner(std::shared_ptr<ts_queue<token>> task, int id, int timeout,
 	auto xx = std::chrono::duration<double>(_now - _last).count(); 
 	std::cout<<xx<<std::endl;
 	if( xx > timeout){
-		struct ipt_counters x = ipt->get_counters();
-		if(end_process(_client, id, x.bcnt, x.pcnt))
-			kill(getpid(), SIGTERM);
+		try{
+			struct ipt_counters x = ipt->get_counters();
+			if(bgp->remove_announce() || bgp->status())
+				if(end_process(_client, id, x.bcnt, x.pcnt))
+					kill(getpid(), SIGTERM);
+			
+		}catch(...){
+			continue;
+		}
 	}
 
         if(task->wait_and_pop(tmp, 1000))
@@ -112,7 +127,8 @@ void task_runner(std::shared_ptr<ts_queue<token>> task, int id, int timeout,
 				else
 					data = str;
 
-				df.insert_record(id,data);
+				if(!df.insert_record(id,data))
+					continue;
 				ipt->add_rule(tmp);
 				rule_list.push_back(str);
 			}
@@ -147,13 +163,16 @@ int main(int argc, char ** argv){
     }
     close(std::atoi(argv[1]));
 
-   using namespace Sqlite;
-   SQLite sq("Taro_Config");
-   auto [bgpid, interface, timeout, bgppass, enable_pass, bgp_ip, bgp_port ] = sq.get_config();
+    using namespace Sqlite;
+    SQLite sq("Taro_Config");
+    auto [bgpid, interface, timeout, bgppass, enable_pass, bgp_ip, bgp_port ] = sq.get_config();
+    if(bgpid == 0){
+	    std::cerr<<"Couldn't get config!";
+		    return 1;
+    }
 
-    std::shared_ptr<Client> _client = std::make_shared<Client>("127.0.0.1", 9200);
-
-
+    boost::asio::io_context io_context_;
+    std::shared_ptr<Client> _client = std::make_shared<Client>(io_context_ ,"127.0.0.1", 9200);
     int id;//, timeout = 5;
     if(input_p[0] == "ID"){
 	    id = std::atoi(input_p[1].c_str());
@@ -165,6 +184,16 @@ int main(int argc, char ** argv){
     b = input_p.begin();
     e = b; e++; e++;
     input_p.erase(b, e);
+
+    std::vector<std::string>::iterator it = std::find(input_p.begin(), input_p.end(), "-d");
+    std::string str = *(++it);
+    uint32_t dst_addr = boost::asio::ip::make_address_v4(str.substr(0, str.find("/"))).to_ulong();
+    std::shared_ptr<Bgp> bgp = std::make_shared<Bgp>(io_context_ ,dst_addr,bgppass, enable_pass, bgp_ip, bgpid, bgp_port);
+    if(!bgp->announce()){
+	    end_process(_client, id, 0,0);
+	    std::cerr<<"Couldn't annouce ip to bgpd."<<std::endl;
+	    return 1;
+    }
 
     std::string _proto = input_p[0];
     input_p.erase(input_p.begin());
@@ -236,7 +265,8 @@ int main(int argc, char ** argv){
     }
     
     threads.add_thread(new boost::thread(watcher, boost::ref(threads_anomly), anomly, task_list));
-    threads.add_thread(new boost::thread(task_runner, task_list, id, timeout, proto_, rule->dst_addr, std::ref(iptable_), std::ref(_client)));
+    threads.add_thread(new boost::thread(task_runner, task_list, id, timeout, proto_, rule->dst_addr,
+			    std::ref(iptable_), std::ref(_client), std::ref(bgp) ));
 
 
     //signals.async_wait([&threads,&io_srv](const boost::system::error_code& e
